Refactor: * split Point from Parser * split downsampling algorithm from Parser * let Plot and Serie be agnostic of real-time format (now the viewer is almsot generic)
